Use the SvgRenderer for Qt 4 as well, if available.
Partial revert of commit 73f285f1480ac34461238b0755973fda09162a8e, which addressed the bug that causes data not to be clipped properly in the svg rendering: the curve is plotted into the top and bottom margins of the plot, outside of the plotting area.
This bug is now considered minor compared to the benefits of getting a SVG rendering. The pixmap rendering is of lower quality.
See issue #54